+++ src/main/java/acs/dao/ProductDao.java
@@ -11,6 +11,6 @@ import java.util.List;
 @Repository
 public interface ProductDao extends PagingAndSortingRepository<ProductEntity, String> {
     List<ProductEntity> findAllByNameLikeIgnoreCase(@Param("name") String name, Pageable pageable);
-    List<ProductEntity> findAllByPriceGreaterThanEqual(@Param("price") Float price, Pageable pageable);
-    List<ProductEntity> findAllByPriceLessThanEqual(@Param("price") Float price, Pageable pageable);
+    List<ProductEntity> findAllByPriceGreaterThanEqual(@Param("price") Double price, Pageable pageable);
+    List<ProductEntity> findAllByPriceLessThanEqual(@Param("price") Double price, Pageable pageable);
 }
